Split `stateSpace` proxy in two and refactor
To simplify proxy logic, there's now a proxy type for each kind of map,
the outer map corresponds to the old `stateSpace` whilst the inner maps
are `weightMap`.
The chain will stringify differently when a token is supposed to be an
object or begin and end tokens, and will double quote strings trying to
mimic the format to prevent generating a different chain on `fromJSON`.
Additionally, many repetitive procedures were refactored into helper
functions.